+++ pkg/app/piped/cloudprovider/kubernetes/diffutil_test.go
